Fix #61: gauss hermite now in pulse with laserbase and docs
